add a nav tag with my name in it
 This command will fail with non-zero status upon error. Some exit codes are:

The section or key is invalid (ret=1),

no section or name was provided (ret=2),

the config file is invalid (ret=3),

the config file cannot be written (ret=4),

you try to unset an option which does not exist (ret=5),

you try to unset/set an option for which multiple lines match (ret=5), or

you try to use an invalid regexp (ret=6).

On success, the command returns the exit code 0.

A list of all available configuration variables can be obtained using the git help --config command.